+++ autopylot/controllers/joystick.py
@@ -1,17 +1,10 @@
-# script based on https://github.com/autorope/donkeycar/blob/dev/donkeycar/parts/controller.py
-
 import array
 import logging
 import os
 import struct
 import threading
 
-from ..utils import memory, utils, settings
-
-try:
-    import keyboard
-except ImportError:
-    logging.warning("not support for keyboard module.")
+from ..utils import memory, utils
 
 
 class Joystick(object):
@@ -213,65 +206,3 @@ class XboxOneJoystick(Joystick):
             }
         else:
             self.memory["controller"] = {}
-
-
-class KeyboardController:
-    """Keyboard controller (debugging purpose)."""
-
-    def __init__(self, memory=memory.mem, *args, **kwargs):
-        self.memory = memory
-
-        self.keys_to_steer = {
-            "q": -1,
-            "z": 0,
-            "d": 1,
-        }
-        self.bkeys = ["s"]
-
-    def update(self):
-        """Update function for the keyboard controller.
-        This function stores in the memory the newly fetched keyboard inputs.
-        """
-        pressed = []
-        bpressed = []
-        for k in self.keys_to_steer.keys():
-            pressed.append(keyboard.is_pressed(k))
-        for bk in self.bkeys:
-            bpressed.append(keyboard.is_pressed(bk))
-
-        sum_steer = 0
-        count = 0
-        for key, k_pressed in zip(self.keys_to_steer.keys(), pressed):
-            if k_pressed:
-                sum_steer += self.keys_to_steer[key]
-                count += 1
-
-        if count != 0:
-            steering = sum_steer / count
-            throttle = 0.5
-
-        else:
-            steering = 0
-            throttle = 0
-
-        if any(bpressed):
-            throttle *= -1
-
-        self.memory["controller"] = {
-            "steering": steering,
-            "throttle": throttle,
-        }
-
-
-def Controller(
-    memory=memory.mem,
-    controller_type=settings.settings.CONTROLLER_TYPE,
-    *args,
-    **kwargs,
-):
-    if controller_type == "xbox":
-        return XboxOneJoystick(memory=memory, *args, **kwargs)
-    elif controller_type == "keyboard":
-        return KeyboardController(memory=memory, *args, **kwargs)
-    else:
-        raise ValueError(f"Unknown controller type {controller_type}")
