+++ main_programs/examples/drive_with_controller.py
@@ -1,25 +1,26 @@
 from autopylot.cameras import camera
-from autopylot.controls import controller, control, state_switcher
-from autopylot.utils import logger, memory
+from autopylot.controls import control
+from autopylot.controllers import controller
+from autopylot.utils import logger, memory, state_switcher
 
 # init the logger handlers
 logger.init()
 
 mem = memory.mem
 
-state = state_switcher.StateSwitcher()
-serial = control.Control()
-cam = camera.Camera()
-js = controller.Controller()
+state_switcher_cls = state_switcher.StateSwitcher()
+control_cls = control.Control()
+camera_cls = camera.Camera()
+controller_cls = controller.Controller()
 
 
 def main():
     while True:
 
-        js.update()  # update joystick
+        controller_cls.update()  # update joystick
 
-        # state.update()  # car state
-        cam.update()  # get the last frame from the camera
+        # state_switcher_cls.update()  # car state
+        camera_cls.update()  # get the last frame from the camera
 
         # if mem["state"] == "stop":
         #     mem["steering"] = 0.0
@@ -40,7 +41,7 @@ def main():
         mem["steering"] = mem["controller"]["steering"]
         mem["throttle"] = mem["controller"]["throttle"]
 
-        serial.update()  # send commands to the memory
+        control_cls.update()
 
 
 if __name__ == "__main__":
